requirements.txt
# requirements.txt for DefectDojo using Python 3.x
asteval==0.9.31
bleach==6.1.0
bleach[css]
celery==5.3.6
coverage==7.4.1
defusedxml==0.7.1
django_celery_results==2.5.1
django-auditlog==2.3.0
django-dbbackup==4.1.0
django-environ==0.11.2
django-filter==23.5
django-imagekit==5.0.0
# This library is very outdated, but is a pillar of DefectDojo
# django-multiselectfield==0.1.12
git+https://github.com/DefectDojo/django-multiselectfield@master#egg=django-multiselectfield
django-polymorphic==3.1.0
django-crispy-forms==2.0
django_extensions==3.2.3
django-slack==5.19.0
# This library is very outdated and not directly. It is used solely for migration
# purposes to django-tagulous, so it must stay
# django-tagging==0.5.0
git+https://github.com/DefectDojo/django-tagging@develop#egg=django-tagging
django-watson==1.6.3
django-prometheus==2.3.1
Django==4.1.13
djangorestframework==3.14.0
html2text==2020.1.16
humanize==4.9.0
jira==3.6.0
PyGithub==1.58.2
lxml==5.1.0
Markdown==3.5.2
mysqlclient==2.1.1
openpyxl==3.1.2
Pillow==10.2.0  # required by django-imagekit
psycopg2-binary==2.9.9
cryptography==42.0.4
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
sqlalchemy==2.0.25  # Required by Celery broker transport
urllib3==1.26.18
uWSGI==2.0.23
vobject==0.9.6.1
whitenoise==5.2.0
titlecase==2.4.1
social-auth-app-django==5.4.1
social-auth-core==4.5.2
Python-jose==3.3.0
gitpython==3.1.41
debugpy==1.8.0
python-gitlab==4.4.0
cpe==1.2.1
packageurl-python==0.13.4
django-crum==0.7.9
JSON-log-formatter==0.5.2
django-split-settings==1.2.0
django-debug-toolbar==4.3.0
django-debug-toolbar-request-history==0.1.4
vcrpy==6.0.1
vcrpy-unittest==0.1.7
django-tagulous==1.3.3
PyJWT==2.8.0
cvss==3.0
django-fieldsignals==0.7.0
hyperlink==21.0.0
django-test-migrations==1.3.0
djangosaml2==1.9.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
django-ratelimit==4.1.0
argon2-cffi==23.1.0
blackduck==1.1.0
pycurl==7.45.2  # Required for Celery Broker AWS (SQS) support
boto3==1.34.35  # Required for Celery Broker AWS (SQS) support
netaddr==0.10.1
vulners==2.1.5
fontawesomefree==6.5.1
